+++ package/example/lib/main.dart
@@ -85,6 +85,7 @@ class _DemoPageState extends State<DemoPage> {
       context: context,
       rootDirectory: rootPath,
       fsType: FilesystemType.file,
+      folderColor: Colors.teal,
       allowedExtensions: ['.txt'],
     );
 
@@ -105,6 +106,7 @@ class _DemoPageState extends State<DemoPage> {
       rootDirectory: rootPath,
       fsType: FilesystemType.folder,
       pickText: 'Save file to this folder',
+      folderColor: Colors.teal,
     );
 
     setState(() {