+++ package/lib/src/filesystem_list.dart
@@ -9,6 +9,7 @@ class FilesystemList extends StatelessWidget {
   final bool isRoot;
   final Directory rootDirectory;
   final FilesystemType fsType;
+  final Color folderColor;
   final List<String> allowedExtensions;
   final ValueChanged<Directory> onChange;
   final ValueSelected onSelect;
@@ -18,6 +19,7 @@ class FilesystemList extends StatelessWidget {
     this.isRoot = false,
     @required this.rootDirectory,
     this.fsType = FilesystemType.all,
+    this.folderColor,
     this.allowedExtensions,
     @required this.onChange,
     @required this.onSelect,
@@ -79,6 +81,7 @@ class FilesystemList extends StatelessWidget {
               return FilesystemListTile(
                 fsType: fsType,
                 item: item,
+                folderColor: folderColor,
                 onChange: onChange,
                 onSelect: onSelect,
               );